Fixing blog tests
